requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
fastapi-jsonrpc
pydantic-settings
